+++ src/LiteBus.Messaging/Internal/Registry/Builders/ErrorHandlerDescriptorBuilder.cs
@@ -23,13 +23,14 @@ internal sealed class ErrorHandlerDescriptorBuilder : IHandlerDescriptorBuilder
         {
             var messageType = @interface.GetGenericArguments()[0];
             var messageResultType = @interface.GetGenericArguments()[1];
+            var tags = handlerType.GetTagsFromAttribute();
 
             yield return new ErrorHandlerDescriptor
             {
                 MessageType = messageType.IsGenericType ? messageType.GetGenericTypeDefinition() : messageType,
                 MessageResultType = messageResultType,
                 Order = order,
-                Tags = ArraySegment<string>.Empty,
+                Tags = tags,
                 HandlerType = handlerType,
             };
         }