+++ tests/LiteBus.Events.UnitTests/UseCases/EventWithTag/EventWithTagEventHandler3.cs
@@ -1,0 +1,13 @@
+using LiteBus.Events.Abstractions;
+
+namespace LiteBus.Events.UnitTests.UseCases.EventWithTag;
+
+public sealed class EventWithTagEventHandler3 : IEventHandler<EventWithTag>
+{
+    public Task HandleAsync(EventWithTag message, CancellationToken cancellationToken = default)
+    {
+        message.ExecutedTypes.Add(GetType());
+
+        return Task.CompletedTask;
+    }
+}